Add files via upload
These are some of my best works from learning Java processing. I have been able to use my knowledge from this language to recreate childhood games that have shaped my career path. 
The television allows you using your keyboard to type and erase, change color of the screen and write as well.
The ping pong allows you to use your mouse and keyboard to move the players in order to play the game.
The robot and fly was one of my beginners project as well as the house.
The fireworks allows you to create fireworks by clicking your mouse.
The  frog hop  uses the keyboard to to move the frog to avoid obstacles.
The snake which was one of my toughest games to recreate uses the keyboard to move. study the code to see its movement instructions.
And the doodle jump uses the right and  left doodle to change positions while using the mouse and/or keyboards to jump. 

I hope you enjoy these games, I hope it reminds you of your childhood as it has for me.


N.B: THIS CODES ONLY RUN ON JAVA PROCESSING!!!